Enable the passing of options for SPI devices with defaults
Adding an options argument to the constructor that is passed into the
SPI constructor.  This allows the end-user to set things like mode,
chipSelect, etc.

Include defaults for the options that are shown to work on the raspberry
pi (see RussTheAerialist/node-spi#17 for background).  15khz was chosen
because it's close to the optimal speed for the mcp3008 according to the
datasheet.